5.1.14 Порефакторит утилитарные функции
