+++ src/components/task-edit.js
@@ -1,6 +1,6 @@
 import AbstractComponent from "./abstract-component.js";
 import {COLORS, DAYS, MONTH_NAMES} from '../const.js';
-import {formatTime} from '../utils.js';
+import {formatTime} from '../utils/common.js';
 
 const createColorsMarkup = (colors, currentColor) => {
   return colors
